CorreÃ§Ã£o de bug
Criado diretÃ³rio progs-int para os programas pequenos escritos na linguagem do IntMUD
Acrescentado arquivo progs-int/txtparaint.int

No IntMUD:
Corrigido bug na funÃ§Ã£o textovar.mudar quando a variÃ¡vel Ã© numÃ©rica e o nÃºmero nÃ£o pode ser representado em um int32

No MUD:
Corrigido bug no comando Renascer
Adicionadas opÃ§Ãµes no acesso via navegador

+++ var-textovar.cpp
@@ -159,7 +159,7 @@ bool TTextoVar::FuncMudar(TVariavel * v)
         switch (sub.TipoVar)
         {
         case TextoVarTipoNum:
-            sub.setInt(v[2].getDouble());
+            sub.setDouble(v[2].getDouble());
             break;
         case TextoVarTipoDec:
             sub.setInt(v[2].getInt());
